Fix IO command reads.
Sometimes, mio gives us a readable event for command FIFO files while we can’t actually read the file. To fix this
problem, I introduced a small loop catching the std::io::ErrorKind::WouldBlock, and simply go on with the loop until
we can read. However, I didn’t realize that this is just a bleep, and that String was actually reading some content;
thus, duplicating the command (that was the “malformed request; trailing character” error).

I also use this commit to remove command splitting, as FIFO are scoped by sessions now, so no risk to have two sessions
writing to same FIFO.