Add support for better asynchronous KTS requests.
KTS can be used in many different ways. We can group them in two main
categories:

- As a standalone server.
- Started from within Kakoune.

When the server is started in standalone, opening Kakoune will always almost
immediately get a reply from the KTS server with the init response — so the
session is initialized so quickly that the user doesn’t have the time to open a
buffer before the session is initialized. However, opening a file with something
like `kak example.rs` will change that race and the buffer is very likely to be
loaded (with all the hooks on it run, especially the important ones for us —
`WinSetOption filetype=(.*)`).

If started from within Kakoune, since the server can take some time to bootstrap
(because it loads grammars and queries), it’s much easier to reproduce the
race issue.

This commit fixes the problem by introducing an “initial” response that
iterates on all the open buffers and set their tree_sitter_lang buffer-option.
Because that is done _right_ after the session is enabled — and we have
installed the hooks to react to setting tree_sitter_lang, only the supported
languages will have the corresponding tree-sitter-buffer-metadata run.

TL;DR: that commit fixes all the problems with asynchronous initialization.

Fixes: https://todo.sr.ht/~hadronized/kak-tree-sitter-features/15
References: https://todo.sr.ht/~hadronized/kak-tree-sitter-bugs/9